requirements.txt
pyrogram==2.0.106
motor==3.3.2
python-dotenv==1.0.0
TgCrypto==1.2.5
